+++ lib/grape/request.rb
@@ -4,6 +4,11 @@ module Grape
 
     alias rack_params params
 
+    def initialize(env, options = {})
+      extend options[:build_params_with] || Grape::Extensions::HashWithIndifferentAccess::ParamBuilder
+      super(env)
+    end
+
     def params
       @params ||= build_params
     end
@@ -14,16 +19,12 @@ module Grape
 
     private
 
-    def build_params
-      params = Hashie::Mash.new(rack_params)
-      if env[Grape::Env::GRAPE_ROUTING_ARGS]
-        args = env[Grape::Env::GRAPE_ROUTING_ARGS].dup
-        # preserve version from query string parameters
-        args.delete(:version)
-        args.delete(:route_info)
-        params.deep_merge!(args)
-      end
-      params
+    def grape_routing_args
+      args = env[Grape::Env::GRAPE_ROUTING_ARGS].dup
+      # preserve version from query string parameters
+      args.delete(:version)
+      args.delete(:route_info)
+      args
     end
 
     def build_headers